Fix section ID.
