+++ assets/src/edit-story/components/library/panes/text/textSets/index.js
@@ -13,56 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * External dependencies
- */
-import { useEffect, useState } from 'react';
-import styled from 'styled-components';
-
-/**
- * WordPress dependencies
- */
-import { __ } from '@wordpress/i18n';
-
-/**
- * Internal dependencies
- */
-import { Section } from '../../../common';
-import { UnitsProvider } from '../../../../../units';
-import { PAGE_RATIO, TEXT_SET_SIZE } from '../../../../../constants';
-import { getTextSets } from './utils';
-import TextSet from './textSet';
-
-const TextSetContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  row-gap: 12px;
-`;
-
-function TextSets() {
-  const [textSets, setTextSets] = useState([]);
-
-  useEffect(() => {
-    getTextSets().then(setTextSets);
-  }, []);
-  const title = __('Text Sets', 'web-stories');
-  return (
-    <Section id={title} title={title}>
-      <TextSetContainer role="group" aria-labelledby={title}>
-        <UnitsProvider
-          pageSize={{
-            width: TEXT_SET_SIZE,
-            height: TEXT_SET_SIZE / PAGE_RATIO,
-          }}
-        >
-          {textSets.map((elements, index) => (
-            <TextSet key={index} elements={elements} />
-          ))}
-        </UnitsProvider>
-      </TextSetContainer>
-    </Section>
-  );
-}
-
-export default TextSets;
+export { default } from './textSets';
